chat_commands 1.1.4
Added norecoil command to grant or remove access to no recoil to a player (toggle)

[DEV]
Introduced the status system to make implementing toggle commands easier.
It allows the dev to get and set a status on a player per command and display messages both for the person running the command and for the player affected when the player status changes (for example no recoil ON/OFF).
Note that the status is always OFF by default since we always grant something to a player